separated checks to own file
